requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
xarray==2024.6.0
h5netcdf==1.7.3
h5py==3.15.1
numpy==1.26.4
pandas==2.2.2
geopandas==0.14.4
pyproj==3.6.1
rasterio==1.3.10
s3fs==2024.6.1
fsspec==2024.6.1
pystac-client==0.7.6
boto3==1.35.15
httpx==0.27.0
matplotlib==3.8.4
orjson==3.10.7
streamlit==1.36.0
jinja2==3.1.4
openpyxl==3.1.5
# optional LLM
openai==1.43.0
fpdf2==2.7.9
